+++ cruft/_commands/utils/cookiecutter.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import re
+import json
 from pathlib import Path
 from typing import Any, Dict, Optional
 from urllib.parse import urlparse
@@ -187,3 +188,11 @@ def generate_cookiecutter_context(
             ) from error
 
     return context
+
+
+def get_extra_context_from_file(extra_context_file: Path) -> Dict[str, Any]:
+    extra_context = {}
+    if extra_context_file.exists():
+        with open(extra_context_file, "r") as f:
+            extra_context = json.load(f)
+    return extra_context