+++ CHANGELOG.md
@@ -17,6 +17,16 @@ OR
 Changelog
 =========
 
+## 2.15.0 - 11 May 2023
+- [Improved #57](https://github.com/cruft/cruft/issues/57): Improve how a project can be updated to use new values of template variables
+
+
+## 2.14.0 - 4 April 2023
+- [Implemented #57](https://github.com/cruft/cruft/issues/57): Accept cookiecutter arguments on the command line for update
+
+## 2.13.0 - 4 April 2023
+- [Implemented #172](https://github.com/cruft/cruft/issues/172): Add support for local extensions
+
 ## 2.12.0 - 28 December 2022
 - Uncapped version requirements (including Python) for greater compatiblity
 - [Fixed issue #166](https://github.com/cruft/cruft/issues/166): Fixed compatiblity with latest version of cookiecutter